+++ baby-gru/src/utils/index.ts
@@ -1,3 +1,3 @@
 export { MoorhenMolecule } from "./MoorhenMolecule";
 export { MoorhenMap } from "./MoorhenMap";
-export { MoleculeRepresentation, RepresentationStyles } from "./MoorhenMoleculeRepresentation";
+export { MoleculeRepresentation, RepresentationStyles } from "./Representation/MoorhenMoleculeRepresentation";
